Project refactored and restructured.
