+++ src/app/auth/auth.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
+import { LayoutModule } from '../layout/layout.module';
 import { AuthDataRoutingModule } from './auth-routing.module';
 import { LoginComponent } from './components/login.component';
 import { effects, reducers } from './store/reducers/index';
@@ -10,6 +11,7 @@ import { effects, reducers } from './store/reducers/index';
  */
 @NgModule({
   imports: [
+    LayoutModule,
     AuthDataRoutingModule,
     StoreModule.forFeature('authdatareducer', reducers),
     EffectsModule.forFeature(effects),
