+++ src/app/sampledata/components/sampledata-alert/sampledata-alert.component.ts
