+++ src/app/sampledata/components/sampledata-grid/sampledata-grid.component.ts
@@ -8,8 +8,8 @@ import { TranslocoService } from '@ngneat/transloco';
 import { Store } from '@ngrx/store';
 import { untilDestroyed } from 'ngx-take-until-destroy';
 import { Observable } from 'rxjs';
-import { AuthService } from '../../../../core/security/auth.service';
-import { Pageable } from '../../../../shared/models/pageable';
+import { AuthService } from '../../../core/security/auth.service';
+import { Pageable } from '../../../shared/models/pageable';
 import { SampleDataDialogComponent } from '../../components/sampledata-dialog/sampledata-dialog.component';
 import { SampleDataModel } from '../../models/sampledata.model';
 import { SearchCriteriaDataModel } from '../../models/searchcriteriadata.model';
