+++ src/app/sampledata/sampledata.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
-import { CoreModule } from '../../core/core.module';
+import { CoreModule } from '../core/core.module';
 import { SampleDataDialogComponent } from '../sampledata/components/sampledata-dialog/sampledata-dialog.component';
 import { SampleDataAlertComponent } from './components/sampledata-alert/sampledata-alert.component';
 import { SampleDataGridComponent } from './components/sampledata-grid/sampledata-grid.component';
