+++ src/app/sampledata/store/reducers/index.ts
